+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `limestone/Card` `splitCaption` prop to allow user to split the content for the captions
+
 ## [5.3.2] - 2025-10-28
 
 No significant changes.